requirements.txt
docx
docxcompose
click
pyyaml
pandoc-acro
pandoc-crossref